Reduce copying on marshalOut and put
Eliminate unnecessary copying by using unsafe ByteString functions.

This is safe for marshalOut because LMDB will not modify the data we provide
it; it is safe for copyLazyBS because we merely copy each ByteString chunk.

It would not be safe for marshalIn because we have no guarantee the value will
be fully decoded before the MDB_val becomes invalid.

By making use of LMDB's MDB_RESERVE feature on 'put', we avoid an unnecessary
toStrict by copying the lazy ByteString directly into LMDB's address space.